--- src/presentation/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './signup.controller'
