+++ src/presentation/controllers/signup/signup-protocols.ts
@@ -1,0 +1,4 @@
+export * from '../../protocols'
+export * from '../../protocols/password-validator'
+export * from '../../protocols/email-validator'
+export * from '../../../domain/usecases/add-account'
