+++ src/presentation/helpers/http-helper.ts
@@ -1,14 +1,14 @@
 import { ServerError } from '../errors'
-import { type IHttpResponse } from '../protocols'
+import { type HttpResponse } from '../protocols'
 
-export function badRequest (error: Error): IHttpResponse {
+export function badRequest (error: Error): HttpResponse {
   return {
     statusCode: 400,
     body: error
   }
 }
 
-export function serverError (): IHttpResponse {
+export function serverError (): HttpResponse {
   return {
     statusCode: 500,
     body: new ServerError()
