+++ src/presentation/protocols/http.ts
@@ -1,8 +1,8 @@
-export interface IHttpRequest {
+export interface HttpRequest {
   body: any
 }
 
-export interface IHttpResponse {
+export interface HttpResponse {
   statusCode: number
   body?: any
 }
